move bit_stream
